pyproject.toml
[project]
name = "foodie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "filterpy>=1.4.5",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "altair>=5.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "numpy>=1.24.0",
    "supabase>=2.25.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.0.0",
    "streamlit-keyup>=0.3.0",
    "requests>=2.31.0",
]
